requirements.txt
# BuddyHelps Voice Server Dependencies
# CRITICAL: cuda-python MUST be 12.6.2 (13.x breaks CUDA graphs = 22x slowdown)

# Core ML
torch>=2.1.0
torchvision
torchaudio

# CUDA (PINNED VERSION - DO NOT CHANGE)
cuda-python==12.6.2

# STT - Parakeet (NVIDIA NeMo)
nemo_toolkit[asr]>=1.23.0

# TTS - Kokoro
kokoro>=0.9.4
soundfile

# LLM - Qwen with vLLM
vllm>=0.4.0
transformers>=4.40.0

# Server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
numpy>=1.26.0

# Audio processing
librosa>=0.10.0
scipy>=1.12.0
